Fix neko server crash on #8
This error was related to a bug on how neko manages synchronous
sockets, which causes the flash policy server to execute endlessly
as an orphan thread, causing some network errors (i.e. port is
already in use).

Also, the method "SocketWrapper#toString" was causing some errors
too on the neko target. This changes fixes it by adding more
conditions.